Started single-image opening. Updated ReadMe & example images
I started adding a feature that allow HDPIC to be used as a subprogram! In layman's terms: it takes an image name stored in Ans and displays that image, then returns to the main program. (It's broken right now though) I also added example images so people can see the amazing difference betweent he beta release and the developer version.